+++ src/Rule/Assertion/Relation/ShouldNotDepend/IncludedTraitsRule.php
@@ -1,0 +1,15 @@
+<?php declare(strict_types=1);
+
+namespace PHPat\Rule\Assertion\Relation\ShouldNotDepend;
+
+use PHPat\Rule\Extractor\Relation\AllTraitsExtractor;
+use PHPStan\Node\InClassNode;
+use PHPStan\Rules\Rule;
+
+/**
+ * @implements Rule<InClassNode>
+ */
+final class IncludedTraitsRule extends ShouldNotDepend implements Rule
+{
+    use AllTraitsExtractor;
+}